mv eval_sample to defaults to rm circular import
